+++ tests/add-fscomp1.sh
@@ -43,6 +43,11 @@ _add_f_to_p()
 	__monitor ADDF2P "$@"
 }
 
+_add_d_to_p()
+{
+	__monitor ADDD2P "$@"
+}
+
 test_pot_add_fscomp_001()
 {
 	pot-add-fscomp
@@ -53,6 +58,7 @@ test_pot_add_fscomp_001()
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 
 	setUp
@@ -64,17 +70,18 @@ test_pot_add_fscomp_001()
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 
 	setUp
 	pot-add-fscomp -b bb
-	assertEquals "Exit rc" "1" "$?"
 	assertEquals "Help calls" "1" "$HELP_CALLS"
 	assertEquals "Error calls" "0" "$ERROR_CALLS"
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 
 	setUp
@@ -86,6 +93,7 @@ test_pot_add_fscomp_001()
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -99,6 +107,7 @@ test_pot_add_fscomp_002()
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 
 	setUp
@@ -110,6 +119,7 @@ test_pot_add_fscomp_002()
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 
 	setUp
@@ -121,17 +131,18 @@ test_pot_add_fscomp_002()
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 
 	setUp
 	pot-add-fscomp -p test-pot -f test-fscomp
-	assertEquals "Exit rc" "1" "$?"
 	assertEquals "Help calls" "1" "$HELP_CALLS"
 	assertEquals "Error calls" "1" "$ERROR_CALLS"
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 
 	setUp
@@ -143,6 +154,7 @@ test_pot_add_fscomp_002()
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 
 	setUp
@@ -154,6 +166,7 @@ test_pot_add_fscomp_002()
 	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -167,6 +180,7 @@ test_pot_add_fscomp_003()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -180,6 +194,7 @@ test_pot_add_fscomp_004()
 	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -193,6 +208,7 @@ test_pot_add_fscomp_005()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -206,6 +222,7 @@ test_pot_add_fscomp_006()
 	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -219,6 +236,7 @@ test_pot_add_fscomp_007()
 	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -232,6 +250,7 @@ test_pot_add_fscomp_008()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -245,6 +264,7 @@ test_pot_add_fscomp_009()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -258,6 +278,63 @@ test_pot_add_fscomp_010()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
 	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
+	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
+}
+
+test_pot_add_fscomp_011()
+{
+	pot-add-fscomp -p test-pot -f zroot/test-fscomp -m /test-mnt -d /tmp
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
+	assertEquals "Error calls" "1" "$ERROR_CALLS"
+	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
+	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
+	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
+	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
+	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
+}
+
+test_pot_add_fscomp_012()
+{
+	pot-add-fscomp -p test-pot -d nodir -m /test-mnt
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
+	assertEquals "Error calls" "1" "$ERROR_CALLS"
+	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
+	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
+	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
+	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
+	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
+}
+
+test_pot_add_fscomp_013()
+{
+	pot-add-fscomp -p test-pot -d .. -m /test-mnt
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
+	assertEquals "Error calls" "1" "$ERROR_CALLS"
+	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
+	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
+	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
+	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
+	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
+}
+
+test_pot_add_fscomp_014()
+{
+	pot-add-fscomp -p test-pot -d /tmp -m /test-mnt -w
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
+	assertEquals "Error calls" "1" "$ERROR_CALLS"
+	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
+	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
+	assertEquals "_is_uid0 calls" "0" "$ISUID0_CALLS"
+	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "0" "$MPVALID_CALLS"
 }
 
@@ -276,6 +353,7 @@ test_pot_add_fscomp_020()
 	assertEquals "_add_f_to_p arg" "/test-mnt" "$ADDF2P_CALL1_ARG3"
 	assertEquals "_add_f_to_p arg" "NO" "$ADDF2P_CALL1_ARG4"
 	assertEquals "_add_f_to_p arg" "" "$ADDF2P_CALL1_ARG5"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "1" "$MPVALID_CALLS"
 
 	setUp
@@ -292,6 +370,7 @@ test_pot_add_fscomp_020()
 	assertEquals "_add_f_to_p arg" "/test-mnt" "$ADDF2P_CALL1_ARG3"
 	assertEquals "_add_f_to_p arg" "external" "$ADDF2P_CALL1_ARG4"
 	assertEquals "_add_f_to_p arg" "" "$ADDF2P_CALL1_ARG5"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "1" "$MPVALID_CALLS"
 }
 
@@ -310,6 +389,7 @@ test_pot_add_fscomp_021()
 	assertEquals "_add_f_to_p arg" "/test-mnt" "$ADDF2P_CALL1_ARG3"
 	assertEquals "_add_f_to_p arg" "NO" "$ADDF2P_CALL1_ARG4"
 	assertEquals "_add_f_to_p arg" "ro" "$ADDF2P_CALL1_ARG5"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "1" "$MPVALID_CALLS"
 
 	setUp
@@ -326,6 +406,7 @@ test_pot_add_fscomp_021()
 	assertEquals "_add_f_to_p arg" "/test-mnt" "$ADDF2P_CALL1_ARG3"
 	assertEquals "_add_f_to_p arg" "external" "$ADDF2P_CALL1_ARG4"
 	assertEquals "_add_f_to_p arg" "ro" "$ADDF2P_CALL1_ARG5"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "1" "$MPVALID_CALLS"
 
 	setUp
@@ -342,6 +423,7 @@ test_pot_add_fscomp_021()
 	assertEquals "_add_f_to_p arg" "/test-mnt" "$ADDF2P_CALL1_ARG3"
 	assertEquals "_add_f_to_p arg" "NO" "$ADDF2P_CALL1_ARG4"
 	assertEquals "_add_f_to_p arg" "zfs-remount" "$ADDF2P_CALL1_ARG5"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
 	assertEquals "_mountpoint_validation calls" "1" "$MPVALID_CALLS"
 
 	setUp
@@ -358,6 +440,60 @@ test_pot_add_fscomp_021()
 	assertEquals "_add_f_to_p arg" "/test-mnt" "$ADDF2P_CALL1_ARG3"
 	assertEquals "_add_f_to_p arg" "NO" "$ADDF2P_CALL1_ARG4"
 	assertEquals "_add_f_to_p arg" "ro" "$ADDF2P_CALL1_ARG5"
+	assertEquals "_add_d_to_p calls" "0" "$ADDD2P_CALLS"
+	assertEquals "_mountpoint_validation calls" "1" "$MPVALID_CALLS"
+}
+
+test_pot_add_fscomp_022()
+{
+	pot-add-fscomp -p test-pot -d /tmp -m /test-mnt
+	assertEquals "Exit rc" "0" "$?"
+	assertEquals "Help calls" "0" "$HELP_CALLS"
+	assertEquals "Error calls" "0" "$ERROR_CALLS"
+	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
+	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
+	assertEquals "_is_uid0 calls" "1" "$ISUID0_CALLS"
+	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "1" "$ADDD2P_CALLS"
+	assertEquals "_add_d_to_p arg" "/tmp" "$ADDD2P_CALL1_ARG1"
+	assertEquals "_add_d_to_p arg" "test-pot" "$ADDD2P_CALL1_ARG2"
+	assertEquals "_add_d_to_p arg" "/test-mnt" "$ADDD2P_CALL1_ARG3"
+	assertEquals "_add_d_to_p arg" "" "$ADDD2P_CALL1_ARG4"
+	assertEquals "_mountpoint_validation calls" "1" "$MPVALID_CALLS"
+
+	setUp
+	pot-add-fscomp -p test-pot -d /tmp -m /test-mnt -e
+	assertEquals "Exit rc" "0" "$?"
+	assertEquals "Help calls" "0" "$HELP_CALLS"
+	assertEquals "Error calls" "0" "$ERROR_CALLS"
+	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
+	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
+	assertEquals "_is_uid0 calls" "1" "$ISUID0_CALLS"
+	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "1" "$ADDD2P_CALLS"
+	assertEquals "_add_d_to_p arg" "/tmp" "$ADDD2P_CALL1_ARG1"
+	assertEquals "_add_d_to_p arg" "test-pot" "$ADDD2P_CALL1_ARG2"
+	assertEquals "_add_d_to_p arg" "/test-mnt" "$ADDD2P_CALL1_ARG3"
+	assertEquals "_add_d_to_p arg" "" "$ADDD2P_CALL1_ARG4"
+	assertEquals "_mountpoint_validation calls" "1" "$MPVALID_CALLS"
+}
+
+test_pot_add_fscomp_023()
+{
+
+	pot-add-fscomp -p test-pot -d /tmp -m /test-mnt -r
+	assertEquals "Exit rc" "0" "$?"
+	assertEquals "Help calls" "0" "$HELP_CALLS"
+	assertEquals "Error calls" "0" "$ERROR_CALLS"
+	assertEquals "_zfs_dataset_valid calls" "0" "$ZDVALID_CALLS"
+	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
+	assertEquals "_is_uid0 calls" "1" "$ISUID0_CALLS"
+	assertEquals "_add_f_to_p calls" "0" "$ADDF2P_CALLS"
+	assertEquals "_add_d_to_p calls" "1" "$ADDD2P_CALLS"
+	assertEquals "_add_d_to_p arg" "/tmp" "$ADDD2P_CALL1_ARG1"
+	assertEquals "_add_d_to_p arg" "test-pot" "$ADDD2P_CALL1_ARG2"
+	assertEquals "_add_d_to_p arg" "/test-mnt" "$ADDD2P_CALL1_ARG3"
+	assertEquals "_add_d_to_p arg" "ro" "$ADDD2P_CALL1_ARG4"
 	assertEquals "_mountpoint_validation calls" "1" "$MPVALID_CALLS"
 }
 
@@ -367,6 +503,7 @@ setUp()
 	ZDVALID_CALLS=0
 	HELP_CALLS=0
 	ADDF2P_CALLS=0
+	ADDD2P_CALLS=0
 	MPVALID_CALLS=0
 	FSVALID_CALLS=0