+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add-fscomp: add option -d, to allow to mount generic directories into a pot (-d and -f are mutual)
+
 ### Changed
 - inherit network: added ipv6 support (automatic)